pyproject.toml
[project]
name = "coop-document-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.124.4",
    "google-generativeai>=0.8.5",
    "jose>=1.0.0",
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.3",
    "msal>=1.34.0",
    "office365-rest-python-client>=2.6.2",
    "openai>=2.9.0",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.1",
    "schedule>=1.2.2",
    "supabase>=2.25.0",
    "tiktoken>=0.12.0",
]
